refactor: array return hook들 타입추론방법 변경
- return [...] as const 와 같은 방식으로 사용
